requirements.txt
docx2pdf==0.1.8
docx2txt==0.8
ffmpeg_python==0.2.0
Flask==3.0.0
pdf2docx==0.5.6
Pillow==10.2.0
pydub==0.25.1
PyPDF2==3.0.1
python_docx==1.1.0
qrcode==7.4.2
reportlab==4.0.8
